pyproject.toml
[tool.poetry]
name = "quickmsg"
version = "0.1.0"
description = ""
authors = ["Artem Sitd  <tym3@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.3"
flask = "^3.0.3"
redis = "^5.0.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
